Made superclass for win and gameover screen. Fixes #173
