+++ src/main/java/net/liquidpineapple/pang/screens/GameOverScreen.java
@@ -1,15 +1,7 @@
 package net.liquidpineapple.pang.screens;
 
-import lombok.Getter;
-
-import net.liquidpineapple.pang.Application;
-import net.liquidpineapple.pang.InputHandler;
-import net.liquidpineapple.pang.gui.NumberToken;
-import net.liquidpineapple.pang.gui.ScoreSystem;
-
-import java.awt.event.KeyEvent;
 import java.io.IOException;
-import java.util.ArrayList;
+
 import javax.imageio.ImageIO;
 
 
@@ -20,60 +12,18 @@ import javax.imageio.ImageIO;
  * @author Jaap-Jan
  * @date 21-9-2016.
  */
-public class GameOverScreen extends Screen {
-  @Getter
-  private static ArrayList<NumberToken> places;
+public class GameOverScreen extends EndGameScreen {
 
   /**
    * Constructor for the game over screen.
    */
   public GameOverScreen() {
+    super();
     try {
       backgroundImage = ImageIO.read(LevelScreen.class.getResource("/sprites/gameover.png"));
     } catch (IOException ex) {
       ex.printStackTrace();
     }
-
-    //These are from right to left.
-    places = new ArrayList<>(9);
-    int numberPosX = 524;
-    int numberPosY = 300;
-    int numberOfTokens = 9;
-    for (int i = 0; i < numberOfTokens; i++) {
-      NumberToken token = new NumberToken(numberPosX, numberPosY);
-      places.add(token);
-      objectList.add(token);
-      numberPosX -= 32;
-    }
-    displayscore();
-  }
-
-  /**
-   * Method that updates the game over screen.
-   */
-  @Override
-  public void doUpdate() {
-    super.doUpdate();
-    if (InputHandler.isKeyPressed(KeyEvent.VK_ENTER)) {
-      Application.lifeSystem.resetLives();
-      Application.getScoreSystem().resetScore();
-      Application.getBoard().changeScreen(new MainMenuScreen());
-    }
-  }
-
-  /**
-   * Method that makes the score display on the screen.
-   */
-  public void displayscore() {
-    int calcscore = ScoreSystem.getScore();
-    int pos = 0;
-
-    while (calcscore > 0) {
-      places.get(pos).setScoreToken(calcscore % 10);
-      calcscore /= 10;
-      pos++;
-    }
-    doUpdate();
   }
 
 }
